backend/requirements.txt
beautifulsoup4 = "4.12.3"
langchain = "0.2.3"
langgraph = "0.0.67"
langchainhub = "0.1.20"
langchain-community = "0.2.4"
black = "24.4.2"
isort = "5.13.2"
tavily-python = "0.3.3"
langchain-chroma = "0.1.1"
python-dotenv = "1.0.1"
pytest = "8.2.2"
langchain-openai = "0.1.8"
fastapi = "0.104.1"
uvicorn = "0.24.0"
python-multipart = "0.0.6"
sqlalchemy = "2.0.23"
psycopg2-binary = "2.9.9"
alembic = "1.12.1"
pyjwt = "2.8.0"
cryptography = "41.0.8"
authlib = "1.2.1"
httpx = "0.25.2"
